refactor(core): centralize FormData header handling
resolveConfig and the HTTP adapter carried separate copies of the formDataHeaderPolicy logic. Move that axios-specific behavior into lib/core/setFormDataHeaders.js so both call sites share the same filtering without expanding the generic helpers layer.

Keep the defensive handling for missing FormData headers, add Node HTTP regression coverage, and record the user-visible robustness fix in the pre-release changelog.
