+++ PRE_RELEASE_CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ## Bug Fixes
 
+- **FormData headers:** Node.js requests using `formDataHeaderPolicy: 'content-only'` now tolerate custom `FormData#getHeaders()` implementations that return no headers instead of throwing while filtering them. (**#11062**)
 - **Data URL size limits:** Corrected base64 `data:` URL size estimation so percent-embedded input cannot bypass `maxContentLength`. The Node HTTP adapter now bounds the raw `Buffer` allocation, including ignored characters and content after padding, while the fetch adapter preserves percent-decoded base64 semantics and excludes URL fragments from the payload estimate. (**#11061**)
 - **AxiosError - aggregate errors:** `AxiosError.from()` now synthesizes a message from nested `AggregateError` entries when the outer message is blank, preserving dual-stack connection failure details in structured logs. (**#11059**, closes **#6721**)
 - **AxiosError:** `AxiosError#toJSON()` now serializes `Set` values in request config snapshots as arrays instead of empty objects. (**#11044**, refs **#5910**)
